+++ src/felupe/constitution/_user_materials_hyperelastic.py
@@ -58,10 +58,17 @@ class UserMaterialHyperelastic(UserMaterial):
         def viscoelastic(C, Cin, mu, eta, dtime):
             "Finite strain viscoelastic material formulation."
 
-            Ci = tm.special.from_triu_1d(Cin) + mu / eta * dtime * det(C) ** (-1 / 3) * C
-            Ci = det(Ci) ** (-1 / 3) * Ci
-            I1 = det(C) ** (-1 / 3) * tm.trace(C @ inv(Ci))
+            # unimodular part of the right Cauchy-Green deformation tensor
+            Cu = det(C) ** (-1 / 3) * C
 
+            # update of state variables by evolution equation
+            Ci = tm.special.from_triu_1d(Cin) + mu / eta * dtime * Cu
+            Ci = det(Ci) ** (-1 / 3) * Ci
+
+            # first invariant of elastic part of right Cauchy-Green deformation tensor
+            I1 = tm.trace(Cu @ inv(Ci))
+
+            # first Piola-Kirchhoff stress tensor and state variable
             return mu / 2 * (I1 - 3), tm.special.triu_1d(Ci)
 
         umat = fem.UserMaterialHyperelastic(
@@ -74,16 +81,16 @@ class UserMaterialHyperelastic(UserMaterial):
     """
 
     def __init__(self, fun, nstatevars=0, parallel=False, **kwargs):
-        
+
         if nstatevars > 0:
             # split the original function into two sub-functions
             self.fun = tr.take(fun, item=0)
             self.fun_statevars = tr.take(fun, item=1)
         else:
             self.fun = fun
-            
+
         self.parallel = parallel
-        
+
         super().__init__(
             stress=self._stress,
             elasticity=self._elasticity,
@@ -93,12 +100,12 @@ class UserMaterialHyperelastic(UserMaterial):
 
     def _stress(self, x, **kwargs):
         F = x[0]
-        
+
         if self.nstatevars > 0:
             statevars = (x[1],)
         else:
             statevars = ()
-            
+
         C = dot(transpose(F), F)
         S = tr.gradient(self.fun, wrt=0, ntrax=2, parallel=self.parallel, sym=True)(
             C, *statevars, **kwargs
@@ -113,12 +120,12 @@ class UserMaterialHyperelastic(UserMaterial):
 
     def _elasticity(self, x, **kwargs):
         F = x[0]
-        
+
         if self.nstatevars > 0:
             statevars = (x[1],)
         else:
             statevars = ()
-            
+
         C = dot(transpose(F), F)
         D, S, W = tr.hessian(
             self.fun, wrt=0, ntrax=2, full_output=True, parallel=self.parallel, sym=True